Move ds_manager structure to ds_manager.h.
